requirements.txt
Django
djangorestframework
django-cors-headers
python-jose[cryptography]
passlib
argon2-cffi
dj-database-url
